core/pyproject.toml
[project]
name = "commons-core"
version = "0.1.0"
description = "Core utilities for Python Commons library"
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Python Commons Team", email = "team@pythoncommons.org"}
]
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-json-logger>=2.0.7",
    "tenacity>=8.2.0",
    "pybreaker>=1.0.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "orjson>=3.9.0",
    "structlog>=23.2.0",
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/commons_core"]